LayoutTests/fast/mediastream/MediaStreamTrackList-expected.txt

Tests MediaStreamTrackList callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Add track callback succeeded.
PASS Remove track callback succeeded.
PASS stream1.videoTracks.add(stream2.videoTracks[0]) threw exception Error: InvalidStateError: DOM Exception 11.
PASS stream1.videoTracks.remove(stream2.videoTracks[0]) threw exception Error: InvalidStateError: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
